Revert "Local Multi 4인 패드 지원 기능 추가."
