+++ EngineSIU/EngineSIU/Engine/Source/Editor/LevelEditor/SlateAppMessageHandler.h
@@ -1,13 +1,10 @@
 #pragma once
-#pragma comment(lib, "xinput.lib")
-
 #include "RawInput.h"
 #include "Delegates/DelegateCombination.h"
 #include "HAL/PlatformType.h"
 #include "InputCore/InputCoreTypes.h"
 #include "Math/Vector.h"
 #include "SlateCore/Input/Events.h"
-#include "xinput.h"
 
 namespace EMouseButtons
 {
@@ -15,21 +12,13 @@ enum Type : uint8;
 }
 
 DECLARE_MULTICAST_DELEGATE_TwoParams(FOnKeyCharDelegate, const TCHAR /*Character*/, const bool /*IsRepeat*/);
-
 DECLARE_MULTICAST_DELEGATE_OneParam(FOnKeyDownDelegate, const FKeyEvent& /*InKeyEvent*/);
 DECLARE_MULTICAST_DELEGATE_OneParam(FOnKeyUpDelegate, const FKeyEvent& /*InKeyEvent*/);
-
 DECLARE_MULTICAST_DELEGATE_OneParam(FOnMouseDownDelegate, const FPointerEvent& /*InMouseEvent*/);
 DECLARE_MULTICAST_DELEGATE_OneParam(FOnMouseUpDelegate, const FPointerEvent& /*InMouseEvent*/);
 DECLARE_MULTICAST_DELEGATE_OneParam(FOnMouseDoubleClickDelegate, const FPointerEvent& /*InMouseEvent*/);
 DECLARE_MULTICAST_DELEGATE_OneParam(FOnMouseWheelDelegate, const FPointerEvent& /*InMouseEvent*/);
 DECLARE_MULTICAST_DELEGATE_OneParam(FOnMouseMoveDelegate, const FPointerEvent& /*InMouseEvent*/);
-
-DECLARE_MULTICAST_DELEGATE_OneParam(FOnControllerButtonDownDelegate, const FControllerButtonEvent& /*InControllerEvent*/);
-DECLARE_MULTICAST_DELEGATE_OneParam(FOnControllerButtonUpDelegate, const FControllerButtonEvent& /*InControllerEvent*/);
-DECLARE_MULTICAST_DELEGATE_OneParam(FOnControllerAnalogDelegate, const FControllerAnalogEvent& /*InControllerEvent*/);
-DECLARE_MULTICAST_DELEGATE_OneParam(FOnControllerConnectedDelegate, uint32);
-DECLARE_MULTICAST_DELEGATE_OneParam(FOnControllerDisconnectedDelegate, uint32);
 
 DECLARE_MULTICAST_DELEGATE_OneParam(FOnRawMouseInputDelegate, const FPointerEvent& /*InRawMouseEvent*/);
 DECLARE_MULTICAST_DELEGATE_OneParam(FOnRawKeyboardInputDelegate, const FKeyEvent& /*InRawKeyboardEvent*/);
@@ -43,8 +32,8 @@ public:
     FSlateAppMessageHandler();
 
     void ProcessMessage(HWND hWnd, uint32 Msg, WPARAM wParam, LPARAM lParam);
-    void UpdateXboxControllers(float DeltaTime);
-    
+
+public:
     /** Cursor와 관련된 변수를 업데이트 합니다. */
     void UpdateCursorPosition(const FVector2D& NewPos);
 
@@ -64,33 +53,14 @@ protected:
     void OnKeyChar(const TCHAR Character, const bool IsRepeat);
     void OnKeyDown(uint32 KeyCode, const uint32 CharacterCode, const bool IsRepeat);
     void OnKeyUp(uint32 KeyCode, const uint32 CharacterCode, const bool IsRepeat);
-    
     void OnMouseDown(const EMouseButtons::Type Button, const FVector2D CursorPos);
     void OnMouseUp(const EMouseButtons::Type Button, const FVector2D CursorPos);
     void OnMouseDoubleClick(const EMouseButtons::Type Button, const FVector2D CursorPos);
     void OnMouseWheel(const float Delta, const FVector2D CursorPos);
     void OnMouseMove();
 
-    // Xbox 컨트롤러 함수
-    bool IsXboxControllerConnected(uint32 ControllerId) const;
-    
-    void OnXboxControllerButtonDown(uint32 ControllerId, EXboxButtons::Type Button, bool bIsRepeat = false);
-    void OnXboxControllerButtonUp(uint32 ControllerId, EXboxButtons::Type Button);
-    void OnXboxControllerAnalogInput(uint32 ControllerId, EXboxAnalog::Type AnalogType, float Value);
-    void OnXboxControllerConnected(uint32 ControllerId);
-    void OnXboxControllerDisconnected(uint32 ControllerId);
-    void SetXboxControllerVibration(uint32 ControllerId, float LeftMotor, float RightMotor);
-
-
-    // Raw Input
     void OnRawMouseInput(const RAWMOUSE& RawMouseInput);
     void OnRawKeyboardInput(const RAWKEYBOARD& RawKeyboardInput);
-
-    void ProcessXboxControllerButtons(uint32 ControllerId);
-    void ProcessXboxControllerAnalogInputs(uint32 ControllerId);
-    
-    float NormalizeThumbstick(SHORT Value, SHORT DeadZone) const;
-    float NormalizeTrigger(BYTE Value) const;
 
     // 추가적인 함수는 UnrealEngine [SlateApplication.h:1628]을 참조
 
@@ -98,19 +68,11 @@ public:
     FOnKeyCharDelegate OnKeyCharDelegate;
     FOnKeyDownDelegate OnKeyDownDelegate;
     FOnKeyUpDelegate OnKeyUpDelegate;
-    
     FOnMouseDownDelegate OnMouseDownDelegate;
     FOnMouseUpDelegate OnMouseUpDelegate;
     FOnMouseDoubleClickDelegate OnMouseDoubleClickDelegate;
     FOnMouseWheelDelegate OnMouseWheelDelegate;
     FOnMouseMoveDelegate OnMouseMoveDelegate;
-
-    // Xbox 컨트롤러 델리게이트들
-    FOnControllerButtonDownDelegate OnControllerButtonDownDelegate;
-    FOnControllerButtonUpDelegate OnControllerButtonUpDelegate;
-    FOnControllerConnectedDelegate OnXboxControllerConnectedDelegate;
-    FOnControllerDisconnectedDelegate OnXboxControllerDisconnectedDelegate;
-    FOnControllerAnalogDelegate OnXboxControllerAnalogDelegate;
 
     FOnRawMouseInputDelegate OnRawMouseInputDelegate;
     FOnRawKeyboardInputDelegate OnRawKeyboardInputDelegate;
@@ -147,12 +109,4 @@ private:
 
 private:
     void HandleRawInput(const RAWINPUT& RawInput);
-
-    // Xbox 컨트롤러 관련 멤버 변수들
-    static constexpr uint32 MaxControllers = 4;
-    XINPUT_STATE XboxControllerStates[MaxControllers];
-    XINPUT_STATE XboxPreviousStates[MaxControllers];
-    bool XboxControllerConnected[MaxControllers];
-    float XboxControllerUpdateTimer;
-    static constexpr float XboxControllerUpdateInterval = 1.0f / 60.0f; // 60Hz 업데이트
 };
