+++ EngineSIU/EngineSIU/Engine/Source/Editor/PropertyEditor/ControlEditorPanel.cpp
@@ -2,7 +2,7 @@
 
 #include "World/World.h"
 
-#include "Actors/EditorPlayer.h"
+#include "Actors/Player.h"
 #include "Actors/LightActor.h"
 #include "Actors/FireballActor.h"
 
@@ -41,8 +41,6 @@
 
 #include "Animation/SkeletalMeshActor.h"
 #include "Engine/Classes/Engine/AssetManager.h"
-#include "GameFramework/Player.h"
-#include "GameFramework/SequencerPlayer.h"
 #include "Particles/ParticleSystemComponent.h"
 
 
@@ -376,7 +374,6 @@ void ControlEditorPanel::CreateModifyButton(const ImVec2 ButtonSize, ImFont* Ico
             { .Label = "CapsuleCol",        .OBJ = OBJ_CAPSULE_COLLISION },
             { .Label = "SkeletalMeshActor", .OBJ = OBJ_SKELETALMESH },
             { .Label = "SequencerPlayer",   .OBJ = OBJ_SEQUENCERPLAYER },
-            { .Label = "Player",   .OBJ = OBJ_PLAYER },
         };
 
         for (const auto& primitive : primitives)
@@ -500,10 +497,6 @@ void ControlEditorPanel::CreateModifyButton(const ImVec2 ButtonSize, ImFont* Ico
                 }
                 case OBJ_CAMERA:
                 case OBJ_PLAYER:
-                {
-                    SpawnedActor = World->SpawnActor<APlayer>();
-                    SpawnedActor->SetActorLabel(TEXT("OBJ_PLAYER"));
-                }
                 case OBJ_END:
                     break;
                 }
