+++ EngineSIU/EngineSIU/Engine/Source/Editor/PropertyEditor/PropertyEditorPanel.cpp
@@ -6,7 +6,7 @@
 //#include <tchar.h>
 
 #include "World/World.h"
-#include "Actors/EditorPlayer.h"
+#include "Actors/Player.h"
 #include "Animation/AnimationAsset.h"
 #include "Animation/AnimSequence.h"
 #include "Animation/AnimSingleNodeInstance.h"
@@ -41,7 +41,6 @@
 #include "imgui/imgui_curve.h"
 #include "Math/Transform.h"
 #include "Animation/AnimStateMachine.h"
-#include "GameFramework/SequencerPlayer.h"
 #include "Particles/ParticleEmitter.h"
 #include "Particles/ParticleSystem.h"
 #include "Particles/ParticleSystemComponent.h"
@@ -328,6 +327,14 @@ void PropertyEditorPanel::RenderForSceneComponent(USceneComponent* SceneComponen
 void PropertyEditorPanel::RenderForCameraComponent(UCameraComponent* InCameraComponent)
 {
     
+}
+
+void PropertyEditorPanel::RenderForPlayerActor(APlayer* InPlayerActor)
+{
+    if (ImGui::Button("SetMainPlayer"))
+    {
+        GEngine->ActiveWorld->SetMainPlayer(InPlayerActor);
+    }
 }
 
 void PropertyEditorPanel::RenderForActor(AActor* SelectedActor, USceneComponent* TargetComponent) const
