+++ EngineSIU/EngineSIU/Engine/Source/Editor/UnrealEd/EditorViewportClient.cpp
@@ -400,7 +400,7 @@ void FEditorViewportClient::GetViewInfo(FMinimalViewInfo& OutViewInfo) const
     bool bGotViewInfo = false;
         
     OutViewInfo = FMinimalViewInfo();
-    if (APlayerController* PC = GEngine->ActiveWorld->GetPlayerController(ViewportIndex))
+    if (APlayerController* PC = GEngine->ActiveWorld->GetPlayerController())
     {
         if (APlayerCameraManager* PCM = PC->PlayerCameraManager)
         {
