--- EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Actors/EditorPlayer.h
@@ -1,56 +1,0 @@
-#pragma once
-#include "Components/SkeletalMeshComponent.h"
-#include "GameFramework/Actor.h"
-#include "UObject/ObjectMacros.h"
-#include "UObject/ObjectTypes.h"
-
-class USkeletalMeshComponent;
-class UCameraComponent;
-class UGizmoBaseComponent;
-class UGizmoArrowComponent;
-class USceneComponent;
-class UPrimitiveComponent;
-class FEditorViewportClient;
-class UStaticMeshComponent;
-
-class AEditorPlayer : public AActor
-{
-    DECLARE_CLASS(AEditorPlayer, AActor)
-
-    AEditorPlayer() = default;
-
-    virtual void Tick(float DeltaTime) override;
-
-    void Input();
-    bool PickGizmo(FVector& RayOrigin, FEditorViewportClient* InActiveViewport);
-    void ProcessGizmoIntersection(UStaticMeshComponent* Component, const FVector& PickPosition, FEditorViewportClient* InActiveViewport, bool& bIsPickedGizmo);
-    void PickActor(const FVector& PickPosition);
-    void AddControlMode();
-    void AddCoordMode();
-    void SetCoordMode(ECoordMode InMode) { CoordMode = InMode; }
-
-private:
-    static int RayIntersectsObject(const FVector& PickPosition, USceneComponent* Component, float& HitDistance, int& IntersectCount);
-    void ScreenToViewSpace(int32 ScreenX, int32 ScreenY, std::shared_ptr<FEditorViewportClient> ActiveViewport, FVector& RayOrigin);
-    void PickedObjControl();
-    void PickedBoneControl();
-    
-    void ControlRotation(USceneComponent* Component, UGizmoBaseComponent* Gizmo, float DeltaX, float DeltaY);
-    
-    void ControlScale(USceneComponent* Component, UGizmoBaseComponent* Gizmo, float DeltaX, float DeltaY);
-    FQuat ControlBoneRotation(FTransform& Component, UGizmoBaseComponent* Gizmo, float DeltaX, float DeltaY);
-    FVector ControlBoneScale(FTransform& Component, UGizmoBaseComponent* Gizmo, float DeltaX, float DeltaY);
-    
-
-    bool bLeftMouseDown = false;
-
-    POINT LastMousePos;
-    EControlMode ControlMode = CM_TRANSLATION;
-    ECoordMode CoordMode = CDM_WORLD;
-    FQuat InitialBoneRotationForGizmo;
-
-public:
-    void SetMode(EControlMode Mode) { ControlMode = Mode; }
-    EControlMode GetControlMode() const { return ControlMode; }
-    ECoordMode GetCoordMode() const { return CoordMode; }
-};
