+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Camera/CameraComponent.h
@@ -12,7 +12,7 @@ public:
     virtual UObject* Duplicate(UObject* InOuter) override;
     virtual void InitializeComponent() override;
     virtual void TickComponent(float DeltaTime) override;
-    void FollowPlayer(int PlayerIndex);
+    void FollowMainPlayer();
 
     float ViewFOV = 70.0f;
     float NearClip = 0.1f;
