+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Camera/PlayerCameraManager.cpp
@@ -3,7 +3,6 @@
 #include "CameraModifier.h"
 #include "GameFramework/PlayerController.h"
 #include "Camera/CameraModifier_CameraShake.h"
-#include "GameFramework/Pawn.h"
 #include "World/World.h"
 
 bool FTViewTarget::Equal(const FTViewTarget& OtherTarget) const
@@ -303,6 +302,7 @@ void APlayerCameraManager::SetViewTarget(class AActor* NewTarget, struct FViewTa
     
 	if (TransitionParams.BlendTime > 0)
 	{
+
 		BlendTimeToGo = TransitionParams.BlendTime;
 
 		AssignViewTarget(PCOwner->GetPossessedActor(), ViewTarget);
