+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/GameFramework/Character.cpp
@@ -1,8 +1,1 @@
 ﻿#include "Character.h"
-
-UObject* ACharacter::Duplicate(UObject* InOuter)
-{
-    ThisClass* NewActor = Cast<ThisClass>(Super::Duplicate(InOuter));
-
-    return NewActor;
-}
