+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/GameFramework/PlayerController.h
@@ -4,7 +4,6 @@
 #include "GameFramework/Actor.h" 
 #include "Classes/Components/InputComponent.h"
 
-class APawn;
 class APlayerCameraManager;
 
 class APlayerController : public AActor
@@ -14,8 +13,6 @@ class APlayerController : public AActor
 public:
     APlayerController() = default;
     virtual ~APlayerController() override = default;
-
-    virtual UObject* Duplicate(UObject* InOuter) override;
 
     virtual void PostSpawnInitialize() override;
 
@@ -31,15 +28,15 @@ public:
 
     UInputComponent* GetInputComponent() const { return InputComponent; }
 
-    void SetViewTarget(AActor* NewViewTarget, FViewTargetTransitionParams TransitionParams);
+    void SetViewTarget(class AActor* NewViewTarget, struct FViewTargetTransitionParams TransitionParams);
 
-    virtual void Possess(APawn* InActor);
+    virtual void Possess(AActor* InActor);
 
     virtual void UnPossess();
     
     virtual void BindAction(const FString& Key, const std::function<void(float)>& Callback);
 
-    APawn* GetPossessedActor() const { return PossessedActor; }
+    AActor* GetPossessedActor() const { return PossessedActor; }
     
     // 카메라 관련 함수
     AActor* GetViewTarget() const;
@@ -57,7 +54,7 @@ protected:
 
     virtual void SetupInputComponent();
 
-    APawn* PossessedActor = nullptr;
+    AActor* PossessedActor = nullptr;
 
     bool bHasPossessed = false;
 };
