+++ EngineSIU/EngineSIU/Engine/Source/Runtime/SlateCore/Input/Events.h
@@ -46,9 +46,6 @@ public:
     /** 키보드 이벤트인지 여부 */
     virtual bool IsKeyEvent() const { return false; }
 
-    /* 컨트롤러 이벤트인지 여부 */
-    virtual bool IsControllerEvent() const { return false; }
-
 public:
     /** 이 키 이벤트가 자동 반복 입력인지 여부를 반환합니다. */
     bool IsRepeat() const
@@ -282,210 +279,4 @@ private:
     const TSet<EKeys::Type>* PressedButtons;
     EKeys::Type EffectingButton;
     float WheelDelta;
-};
-
-// Xbox 컨트롤러 버튼 이벤트 구조체
-struct FControllerButtonEvent : public FInputEvent
-{
-    FControllerButtonEvent()
-        : FInputEvent(FModifierKeysState{}, IE_None)
-        , Button(EXboxButtons::Invalid)
-        , ControllerId(0)
-        , bIsRepeat(false)
-    {
-    }
-
-    FControllerButtonEvent(
-        const EXboxButtons::Type InButton,
-        const FModifierKeysState& InModifierKeys,
-        EInputEvent InInputEvent,
-        const uint32 InControllerId,
-        const bool bInIsRepeat = false
-    )
-        : FInputEvent(InModifierKeys, InInputEvent)
-        , Button(InButton)
-        , ControllerId(InControllerId)
-        , bIsRepeat(bInIsRepeat)
-    {
-    }
-
-public:
-    EXboxButtons::Type GetButton() const
-    {
-        return Button;
-    }
-
-    uint32 GetControllerId() const
-    {
-        return ControllerId;
-    }
-
-    bool IsRepeat() const
-    {
-        return bIsRepeat;
-    }
-
-    virtual bool IsControllerEvent() const override { return true; }
-
-private:
-    EXboxButtons::Type Button;
-    uint32 ControllerId;
-    bool bIsRepeat;
-};
-
-// Xbox 컨트롤러 아날로그 이벤트 구조체
-struct FControllerAnalogEvent : public FInputEvent
-{
-    FControllerAnalogEvent()
-        : FInputEvent(FModifierKeysState{}, IE_None)
-        , AnalogType(EXboxAnalog::Invalid)
-        , AnalogValue(0.0f)
-        , ControllerId(0)
-        , DeltaTime(0.0f)
-    {
-    }
-
-    FControllerAnalogEvent(
-        const EXboxAnalog::Type InAnalogType,
-        const FModifierKeysState& InModifierKeys,
-        EInputEvent InInputEvent,
-        const float InAnalogValue,
-        const uint32 InControllerId,
-        const float InDeltaTime = 0.0f
-    )
-        : FInputEvent(InModifierKeys, InInputEvent)
-        , AnalogType(InAnalogType)
-        , AnalogValue(InAnalogValue)
-        , ControllerId(InControllerId)
-        , DeltaTime(InDeltaTime)
-    {
-    }
-
-public:
-    EXboxAnalog::Type GetAnalogType() const
-    {
-        return AnalogType;
-    }
-
-    float GetAnalogValue() const
-    {
-        return AnalogValue;
-    }
-
-    uint32 GetControllerId() const
-    {
-        return ControllerId;
-    }
-
-    float GetDeltaTime() const
-    {
-        return DeltaTime;
-    }
-
-    virtual bool IsKeyEvent() const override { return false; }
-    virtual bool IsControllerEvent() const override { return true; }
-    virtual bool IsAnalogEvent() const { return true; }
-
-private:
-    EXboxAnalog::Type AnalogType;
-    float AnalogValue;
-    uint32 ControllerId;
-    float DeltaTime;
-};
-
-// 통합 Xbox 컨트롤러 이벤트 구조체
-struct FXboxControllerEvent : public FInputEvent
-{
-    FXboxControllerEvent()
-        : FInputEvent(FModifierKeysState{}, IE_None)
-        , Button(EXboxButtons::Invalid)
-        , AnalogType(EXboxAnalog::Invalid)
-        , AnalogValue(0.0f)
-        , ControllerId(0)
-        , bIsRepeat(false)
-        , bIsAnalogEvent(false)
-    {
-    }
-
-    // 버튼 이벤트용 생성자
-    FXboxControllerEvent(
-        const EXboxButtons::Type InButton,
-        const FModifierKeysState& InModifierKeys,
-        EInputEvent InInputEvent,
-        const uint32 InControllerId,
-        const bool bInIsRepeat = false
-    )
-        : FInputEvent(InModifierKeys, InInputEvent)
-        , Button(InButton)
-        , AnalogType(EXboxAnalog::Invalid)
-        , AnalogValue(0.0f)
-        , ControllerId(InControllerId)
-        , bIsRepeat(bInIsRepeat)
-        , bIsAnalogEvent(false)
-    {
-    }
-
-    // 아날로그 이벤트용 생성자
-    FXboxControllerEvent(
-        const EXboxAnalog::Type InAnalogType,
-        const FModifierKeysState& InModifierKeys,
-        EInputEvent InInputEvent,
-        const float InAnalogValue,
-        const uint32 InControllerId
-    )
-        : FInputEvent(InModifierKeys, InInputEvent)
-        , Button(EXboxButtons::Invalid)
-        , AnalogType(InAnalogType)
-        , AnalogValue(InAnalogValue)
-        , ControllerId(InControllerId)
-        , bIsRepeat(false)
-        , bIsAnalogEvent(true)
-    {
-    }
-
-public:
-    EXboxButtons::Type GetButton() const
-    {
-        return Button;
-    }
-
-    EXboxAnalog::Type GetAnalogType() const
-    {
-        return AnalogType;
-    }
-
-    float GetAnalogValue() const
-    {
-        return AnalogValue;
-    }
-
-    uint32 GetControllerId() const
-    {
-        return ControllerId;
-    }
-
-    bool IsRepeat() const
-    {
-        return bIsRepeat;
-    }
-
-    bool IsAnalogInput() const
-    {
-        return bIsAnalogEvent;
-    }
-
-    bool IsButtonInput() const
-    {
-        return !bIsAnalogEvent;
-    }
-
-    virtual bool IsControllerEvent() const override { return true; }
-
-private:
-    EXboxButtons::Type Button;
-    EXboxAnalog::Type AnalogType;
-    float AnalogValue;
-    uint32 ControllerId;
-    bool bIsRepeat;
-    bool bIsAnalogEvent;
 };
