+++ EngineSIU/EngineSIU/Shaders/CommonMeshPixelShader.hlsl
@@ -151,7 +151,7 @@ float4 mainPS(PS_INPUT_CommonMesh Input) : SV_Target
         );
 
         // Apply Emissive
-        float3 FinalRGB = LitResult.rgb /*+ EmissiveColor*/;
+        float3 FinalRGB = LitResult.rgb + EmissiveColor;
         float FinalA = LitResult.a;
         
         FinalPixelColor = float4(FinalRGB, FinalA);
@@ -159,7 +159,7 @@ float4 mainPS(PS_INPUT_CommonMesh Input) : SV_Target
     }
     else
     {
-        float3 UnlitRGB = DiffuseColor/* + EmissiveColor*/;
+        float3 UnlitRGB = DiffuseColor + EmissiveColor;
 
         FinalPixelColor = float4(UnlitRGB, BaseAlpha);
     }
